+++ .eslintrc.js
@@ -1,6 +1,7 @@
-const { typescript: { rules } } = require('@zero-plusplus/eslint-my-rules');
+const { typescript } = require('@zero-plusplus/eslint-my-rules');
 
 module.exports = {
+  root: true,
   parser: '@typescript-eslint/parser',
   parserOptions: {
     ecmaVersion: 2018,
@@ -15,21 +16,6 @@ module.exports = {
   },
   plugins: [ "@typescript-eslint" ],
   rules: {
-    ...rules,
-    'arrow-body-style': 'off',
-    'class-methods-use-this': 'off',
-    'lines-between-class-members': 'off',
-    'newline-per-chained-call': ["error", { "ignoreChainWithDepth": 5 }],
-    'no-plusplus': 'off',
-    'no-shadow': 'off',
-    'prefer-named-capture-group': 'off',
-    "prefer-destructuring": 'off',
-    '@typescript-eslint/no-use-before-define': 'off',
-    '@typescript-eslint/no-non-null-assertion': 'off',
-    '@typescript-eslint/no-extra-parens': [ 'error', 'all', { 'enforceForArrowConditionals': false }],
-    '@typescript-eslint/no-type-alias': [ 'error', {
-      allowAliases: 'always',
-      allowCallbacks: 'always',
-    } ],
+    ...typescript.rules,
   }
 }